Add block styles before presets classes.
This is because we want the source order specificity for presets
to be higher than for block styles, to fix cases in which the
block selector has the same specificity as the class.

Ex:

- an user attaches the .has-white-color class to a block
  of type "my block", whose selector is .wp-block-my-block.
- the themes targets the color of "my block" as well,
  generating a rule with the .wp-block-my-block selector.

The color applied to the block that was tweaked by the user
will be the last in the source.